Added Apache 2.0 license file
The CRIU GO bindings are licensed under the Apache 2.0 license. The code
is based on the CRIU repository which is GPLv2 or LGPL. To be able to
relicense the GPLv2 and LGPL code under Apache 2.0 all authors have
agreed. The code in this repository is based on contributions from

 * Jacob Wen <jian.w.wen@oracle.com>
 * Andrei Vagin <avagin@virtuozzo.com>
 * Pavel Emelyanov <xemul@virtuozzo.com>

and all have agreed to relicense it under Apache 2.0.

https://lists.openvz.org/pipermail/criu/2018-October/042809.html
https://lists.openvz.org/pipermail/criu/2018-October/042810.html
https://lists.openvz.org/pipermail/criu/2018-October/042811.html

Signed-off-by: Adrian Reber <areber@redhat.com>