Do not pass rpc.CriuOpts by value, bump to v5
rpc.CriuOpts contains a mutex and therefore should not ever be passed by
value (the other reason is it probably uses too much stack space --
though I haven't checked that).

This fixes a bunch of warnings from govet and copylocks linters, like:

> main.go:190:26: copylocks: Dump passes lock by value: github.com/checkpoint-restore/go-criu/v4/rpc.CriuOpts contains google.golang.org/protobuf/internal/impl.MessageState contains sync.Mutex (govet)
> func (c *Criu) Dump(opts rpc.CriuOpts, nfy Notify) error {
>                          ^
> main.go:195:29: copylocks: Restore passes lock by value: github.com/checkpoint-restore/go-criu/v4/rpc.CriuOpts contains google.golang.org/protobuf/internal/impl.MessageState contains sync.Mutex (govet)
> func (c *Criu) Restore(opts rpc.CriuOpts, nfy Notify) error {
>                             ^

Fix the prototypes, and adjust all the callers accordingly.

Finally, as this changes the public API, bump the version to v5.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>