Switch to google.golang.org/protobuf
This fixes warnings like this one:

> main.go:12:2: SA1019: package github.com/golang/protobuf/proto is deprecated: Use the "google.golang.org/protobuf/proto" package instead. (staticcheck)
> 	"github.com/golang/protobuf/proto"
> 	^

(and there's no functional or API change as far as I can see).

Note that github.org/google/protobuf is still there in go.mod
as it is used by the .pb.go files. Apparently it is there only to ensure
backward compatibility with something like:

> import proto "github.com/golang/protobuf/proto"
> // This is a compile-time assertion that a sufficiently up-to-date version
> // of the legacy proto package is being used.
> const _ = proto.ProtoPackageIsVersion4

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>