deprecate: replace stats and rpc with crit
The `crit` library allows users to decode all image files, including
stats-dump and stats-restore, which is currently being done by the
`stats` package. It also provides the proto bindings present in the
`rpc` package. Hence, `stats` and `rpc` have been deprecated in favour
of `crit` and will be removed in v6.0.0.

Signed-off-by: Prajwal S N <prajwalnadig21@gmail.com>