crit(proto-gen): move to Python script
Previously, CRIT was using an overengineered Makefile stunt to generate
protobuf bindings for the CLI. The problem of statically linking all the
bindings while using the library as a dependency cannot be resolved in a
reasonable manner in the Makefile. The logic has been moved out into a
Python script.

Signed-off-by: Prajwal S N <prajwalnadig21@gmail.com>