ci: cancel preceding workflows run
Add a concurrency group to the main CI workflow to automatically cancel
any in-progress workflows when a pull request has been updated.

This `concurrency` group ensures that a single workflow will run at a
time. For example, when a pull request is updated with a force-push,
the GiHub CI workflows currently in-progress will be automatically
cancelled, and the CI would run only with the updated commits.

https://docs.github.com/en/actions/using-jobs/using-concurrency

An exception is used for the 'master' branch, so that when multiple
pull requests have been merged, the CI would run for each merge commit.

Co-authored-by: Prajwal S N <prajwalnadig21@gmail.com>
Signed-off-by: Radostin Stoyanov <rstoyanov@fedoraproject.org>