crit: run unit tests with coverage results
There are two ways we can enable test coverage for the unit tests:
1) `-coverprofile` and 2) `-cover`. The first method stores the results
in a file that we can upload with `codecov`. The second method only
outputs the test coverage results to standard output.

This patch set the default value of GOFLAGS to `-cover` to
always show the coverage of unit tests.

Suggested-by: Adrian Reber <areber@redhat.com>
Signed-off-by: Radostin Stoyanov <rstoyanov@fedoraproject.org>