refactor(crit): pass proto struct as argument
The proto struct required for encoding or decoding the images was
previously being inferred from the `images.ProtoHandler()` function.
Now, it is inferred using that only in the CLI, and is passed to the
actual CRIT functions as a parameter. This eliminates the redundant
protobuf bindings being imported when using CRIT as a library.

Signed-off-by: Prajwal S N <prajwalnadig21@gmail.com>