test: add `mmapper` test program
This commit introduces a new test program called `mmapper`. `mmapper` is a
simple C program that accepts three arguments: the memory size,  mmap flags
(a combinaison of `s` for `MAP_SHARED`, `p` for  `MAP_PRIVATE`, and optionally
`a` for `MAP_ANONYMOUS`). It then creates a child process with the corresponding
memory mapping of the specified size. This program is used for testing the
accuracy of the `GetShmemSize()` method of `MemoryReader`struct.

Signed-off-by: Kouame Behouba Manasse <behouba@gmail.com>