feat(crit): add SearchPattern method on MemoryReader
This commit adds a new method `SearchPattern` to `MemoryReader` to search
for patterns inside the process memory pages. This method accept regular
expressions for flexible pattern matching, a context (number of bytes
before and after the pattern match), and a size of memory chunk to be
read at a time.

Signed-off-by: Kouame Behouba Manasse <behouba@gmail.com>