ci: add codecov workflow
Uploading code coverage currently fails with the following error:

	Rate limit reached. Please upload with the Codecov repository
	upload token to resolve issue.

This patch replaces the codecov Makefile target, that was used to upload
the coverage files, with a GitHub CI workflow [1]. This change allows us
to specify a Codecov token [2] to authorize report uploads.

In addition, this patch updates the version of Go used for coverage to
1.22 and configures the CI build to fail if Codecov encounters an error
during upload.

[1] https://github.com/codecov/codecov-action
[2] https://docs.codecov.com/docs/github-2-getting-a-codecov-account-and-uploading-coverage

Signed-off-by: Radostin Stoyanov <rstoyanov@fedoraproject.org>